refactor: move GetTLSConfig
