+++ pkg/cache/client.go
@@ -1,9 +1,6 @@
 package cache
 
 import (
-	"crypto/tls"
-	"crypto/x509"
-	"os"
 	"strings"
 	"time"
 
@@ -65,22 +62,4 @@ func New(opts Options) (Cache, func(), error) {
 		return nil, cleanup, xerrors.Errorf("unknown cache type: %s", t)
 	}
 	return cache, func() { _ = cache.Close() }, nil
-}
-
-// GetTLSConfig gets tls config from CA, Cert and Key file
-func GetTLSConfig(caCertPath, certPath, keyPath string) (*x509.CertPool, tls.Certificate, error) {
-	cert, err := tls.LoadX509KeyPair(certPath, keyPath)
-	if err != nil {
-		return nil, tls.Certificate{}, err
-	}
-
-	caCert, err := os.ReadFile(caCertPath)
-	if err != nil {
-		return nil, tls.Certificate{}, err
-	}
-
-	caCertPool := x509.NewCertPool()
-	caCertPool.AppendCertsFromPEM(caCert)
-
-	return caCertPool, cert, nil
 }
